reorganized the _ned_ module
