+++ citation_extractor/ned/__init__.py
@@ -1,0 +1,8 @@
+"""TODO."""
+
+NIL_URN = 'urn:cts:GreekLatinLit:NIL'
+LANGUAGES = ['en', 'es', 'de', 'fr', 'it']
+PREPS = [u'di', u'da', u'of', u'von', u'de']
+AUTHOR_TYPE = 'AAUTHOR'
+WORK_TYPE = 'AWORK'
+REFAUWORK_TYPE = 'REFAUWORK'
